Refactor phase
I've refactored the interfaces, splitting them up into
a WritableStorage and a ReadableStorage. This makes
it possible to complete the ChunkStorage without having
to think about reading the files.

The File class has been renamed to "Source", which is just
and InputStream with metadata. It will be accompanied by
a, guess what, yes Sink in the future (outputstream with meta).